requirements-2.txt
aiofiles
aiohttp==3.9.3
aiosignal==1.3.1
anyio>=3.5.0, <4  
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.2
certifi==2024.2.2
chardet==5.1.0
charset-normalizer==3.3.2
click==8.1.3
python-dotenv==1.0.0
deepgram-sdk==3.2.1
fastapi==0.110.0
frozenlist==1.4.1
groq==0.4.2
gunicorn==20.1.0
h11==0.14.0
idna==3.6
Jinja2==3.1.2
langchain==0.1.12
langchain-community==0.0.28
langchain-groq==0.0.1
lxml==4.9.2
MarkupSafe==2.1.2
multidict==6.0.5
mysql-connector-python==8.0.33
nest-asyncio==1.5.6
nltk==3.8.1
openai==1.55.3
pandas==1.5.3
pinecone-client==3.1.0
pydantic==2.6.4
pytz==2023.3
regex==2023.3.23
requests==2.31.0
SQLAlchemy==2.0.28
starlette==0.36.3
tenacity==8.2.3
tiktoken==0.3.3
tldextract==3.4.4
typing_extensions>=4.9.0, <5.0.0
urllib3==2.2.1
uvicorn==0.20.0
watchfiles==0.19.0
yarl==1.9.4
python-multipart
